500에러, keyArray에서 R0  export를 제공하지 않는다고 뜸(R0이 없는데)
