+++ utils/keyArray.ts
@@ -1,76 +1,75 @@
-const R0: string[] = [
-    "ControlLeft",
-    "MetaLeft",
-    "AltLeft",
-    "Space",
-    "AltRight",
-    "MetaRight",
-    "Fn",
-    "ControlRight",
+const keyArray: string[][] = [
+    [
+        "ControlLeft",
+        "MetaLeft",
+        "AltLeft",
+        "Space",
+        "AltRight",
+        "MetaRight",
+        "Fn",
+        "ControlRight",
+    ],
+    [
+        "ShiftLeft",
+        "KeyZ",
+        "KeyX",
+        "KeyC",
+        "KeyV",
+        "KeyB",
+        "KeyN",
+        "KeyM",
+        "Comma",
+        "Period",
+        "Slash",
+        "ShiftRight",
+    ],
+    [
+        "CapsLock",
+        "KeyA",
+        "KeyS",
+        "KeyD",
+        "KeyF",
+        "KeyG",
+        "KeyH",
+        "KeyJ",
+        "KeyK",
+        "KeyL",
+        "Semicolon",
+        "Quote",
+        "Enter",
+    ],
+    [
+        "Tab",
+        "KeyQ",
+        "KeyW",
+        "KeyE",
+        "KeyR",
+        "KeyT",
+        "KeyY",
+        "KeyU",
+        "KeyI",
+        "KeyO",
+        "KeyP",
+        "BracketLeft",
+        "BracketRight",
+        "Backslash",
+    ],
+    [
+        "Escape",
+        "Digit1",
+        "Digit2",
+        "Digit3",
+        "Digit4",
+        "Digit5",
+        "Digit6",
+        "Digit7",
+        "Digit8",
+        "Digit9",
+        "Digit0",
+        "Minus",
+        "Equal",
+        "Backspace",
+    ],
 ]
 
-const R1: string[] = [
-    "ShiftLeft",
-    "KeyZ",
-    "KeyX",
-    "KeyC",
-    "KeyV",
-    "KeyB",
-    "KeyN",
-    "KeyM",
-    "Comma",
-    "Period",
-    "Slash",
-    "ShiftRight",
-]
-
-const R2: string[] = [
-    "CapsLock",
-    "KeyA",
-    "KeyS",
-    "KeyD",
-    "KeyF",
-    "KeyG",
-    "KeyH",
-    "KeyJ",
-    "KeyK",
-    "KeyL",
-    "Semicolon",
-    "Quote",
-    "Enter",
-]
-
-const R3: string[] = [
-    "Tab",
-    "KeyQ",
-    "KeyW",
-    "KeyE",
-    "KeyR",
-    "KeyT",
-    "KeyY",
-    "KeyU",
-    "KeyI",
-    "KeyO",
-    "KeyP",
-    "BracketLeft",
-    "BracketRight",
-    "Backslash",
-]
-
-const R4: string[] = [
-    "Escape",
-    "Digit1",
-    "Digit2",
-    "Digit3",
-    "Digit4",
-    "Digit5",
-    "Digit6",
-    "Digit7",
-    "Digit8",
-    "Digit9",
-    "Digit0",
-    "Minus",
-    "Equal",
-    "Backspace",
-]
-export { R0, R1, R2, R3, R4 }
+export default keyArray
